Removed Old Dev UI: Arc
