+++ extensions/arc/deployment/src/main/java/io/quarkus/arc/deployment/devui/ArcDevUIProcessor.java
@@ -12,11 +12,6 @@ import io.quarkus.arc.deployment.AnnotationsTransformerBuildItem;
 import io.quarkus.arc.deployment.ArcConfig;
 import io.quarkus.arc.deployment.BeanDefiningAnnotationBuildItem;
 import io.quarkus.arc.deployment.CustomScopeAnnotationsBuildItem;
-import io.quarkus.arc.deployment.devconsole.DevBeanInfo;
-import io.quarkus.arc.deployment.devconsole.DevBeanInfos;
-import io.quarkus.arc.deployment.devconsole.DevDecoratorInfo;
-import io.quarkus.arc.deployment.devconsole.DevInterceptorInfo;
-import io.quarkus.arc.deployment.devconsole.DevObserverInfo;
 import io.quarkus.arc.processor.AnnotationsTransformer;
 import io.quarkus.arc.runtime.devconsole.InvocationInterceptor;
 import io.quarkus.arc.runtime.devconsole.InvocationTree;
