+++ extensions/arc/deployment/src/main/java/io/quarkus/arc/deployment/devui/DevBeanInfo.java
@@ -1,4 +1,4 @@
-package io.quarkus.arc.deployment.devconsole;
+package io.quarkus.arc.deployment.devui;
 
 import java.util.ArrayList;
 import java.util.HashSet;
