+++ extensions/arc/deployment/src/main/java/io/quarkus/arc/deployment/devui/Name.java
@@ -1,4 +1,4 @@
-package io.quarkus.arc.deployment.devconsole;
+package io.quarkus.arc.deployment.devui;
 
 import java.util.Iterator;
 import java.util.List;
