--- extensions/arc/deployment/src/main/resources/dev-templates/beans.html
@@ -1,82 +1,0 @@
-{#include main fluid=true}
-  {#style}
-   .annotation {
-     color: gray;
-     font-style: italic;
-   }
-   span.larger-badge {
-     font-size: 0.9em;
-   }
-   span.app-class {
-     cursor:pointer;
-     color:blue;
-     text-decoration:underline;
-   }
-
-  {/style}
-
-  {#script}
-    $(document).ready(function(){
-      if (!ideKnown()) {
-        return;
-      }
-      $(".class-candidate").each(function() {
-        var className = $(this).text();
-        if (appClassLocation(className)) {
-          $(this).addClass("app-class");
-        }
-      });
-
-      $(".app-class").on("click", function() {
-        openInIDE($(this).text());
-      });
-    });
-  {/script}
-
-  {#title}Beans{/title}
-  {#body}
-  <div class="alert alert-primary alert-dismissible fade show" role="alert" data-timer="30000">
-    Beans are sorted by the bean class name in ascending order. However, application beans go first. 
-    <button type="button" class="close" data-dismiss="alert" aria-label="Close">
-        <span aria-hidden="true">&times;</span>
-    </button>
-  </div>
-  <table class="table table-striped">
-   <thead class="thead-dark">
-    <tr>
-      <th scope="col">#</th>
-      <th scope="col">Bean</th>
-      <th scope="col">Kind</th>
-      <th scope="col">Associated Interceptors</th>
-      <th scope="col">Actions</th>
-    </tr>
-   </thead>
-   <tbody>
-    {#for bean in info:devBeanInfos.beans}
-    <tr>
-      <td>{bean_count}</td>
-      <td>
-       {#display-bean bean/}
-      </td>
-      <td>
-       {#bean-declaration bean/}
-      </td>
-      <td>
-       <ul>
-       {#for interceptorId in bean.interceptors}
-         {#set interceptor=info:devBeanInfos.getInterceptor(interceptorId)}
-         <li><span class="class-candidate">{interceptor.interceptorClass}</span> <span class="badge rounded-pill bg-info text-light larger-badge" title="Priority: {interceptor.priority}">{interceptor.priority}</span></li>
-         {/set}
-       {/for}
-       </ul>
-      </td>
-      <td>
-        {#if !info:devBeanInfos.getDependencyGraph(bean.id).links.empty}
-        <a class="btn btn-primary btn-sm" href="beanDependencyGraph?beanId={bean.id}" role="button" title="Show dependency graph"><i class="fa-solid fa-diagram-project"></i></a>
-        {/if}
-      </td>
-    {/for}
-   </tbody>
-  </table>
-  {/body}
-{/include}
