--- extensions/arc/deployment/src/main/resources/dev-templates/embedded.html
@@ -1,31 +1,0 @@
-<a href="{urlbase}/beans" class="badge badge-light">
-  <i class="fa fa-egg fa-fw"></i>
-  Beans <span class="badge badge-light">{info:devBeanInfos.beans.size}</span></a>
-<br>
-<a href="{urlbase}/observers" class="badge badge-light">
-  <i class="fa fa-eye fa-fw"></i>
-  Observers <span class="badge badge-light">{info:devBeanInfos.observers.size}</span></a>
-<br>
-<a href="{urlbase}/interceptors" class="badge badge-light">
-  <i class="fa fa-traffic-light fa-fw"></i>
-  Interceptors <span class="badge badge-light">{info:devBeanInfos.interceptors.size}</span></a>
-<br>
-{#if !info:devBeanInfos.decorators.empty}
-<a href="{urlbase}/decorators" class="badge badge-light">
-  <i class="fa fa-traffic-light fa-fw"></i>
-  Decorators <span class="badge badge-light">{info:devBeanInfos.decorators.size}</span></a>
-<br>
-{/if}
-{#if config:property('quarkus.arc.dev-mode.monitoring-enabled') is "true"}
-<a href="{urlbase}/events" class="badge badge-light">
-  <i class="far fa-eye fa-fw"></i>
-  Fired Events</a>
-<br>
-<a href="{urlbase}/invocations" class="badge badge-light">
-  <i class="fas fa-project-diagram fa-fw"></i>
-  Invocation Trees</a>
-<br>
-{/if}
-<a href="{urlbase}/removed-beans" class="badge badge-light">
-  <i class="fa fa-trash-alt fa-fw"></i>
-  Removed Components <span class="badge badge-light">{info:devBeanInfos.removedComponents}</span></a>
