--- extensions/arc/deployment/src/main/resources/dev-templates/interceptors.html
@@ -1,76 +1,0 @@
-{#include main fluid=true}
-  {#style}
-   .annotation {
-     color: gray;
-     font-style: italic;
-   }
-   span.larger-badge {
-     font-size: 0.9em;
-   }
-   span.app-class {
-     cursor:pointer;
-     color:blue;
-     text-decoration:underline;
-   }
-  {/style}
-
-  {#script}
-    $(document).ready(function(){
-      if (!ideKnown()) {
-        return;
-      }
-      $(".class-candidate").each(function() {
-        var className = $(this).text();
-        if (appClassLocation(className)) {
-          $(this).addClass("app-class");
-        }
-      });
-
-      $(".app-class").on("click", function() {
-        openInIDE($(this).text());
-      });
-    });
-
-  {/script}
-
-  {#title}Interceptors{/title}
-  {#body}
-  <div class="alert alert-primary alert-dismissible fade show" role="alert" data-timer="30000">
-    Interceptors are sorted by the interceptor class name in ascending order. However, application interceptors go first. 
-    <button type="button" class="close" data-dismiss="alert" aria-label="Close">
-        <span aria-hidden="true">&times;</span>
-    </button>
-  </div>
-  <table class="table table-striped">
-   <thead class="thead-dark">
-    <tr>
-      <th scope="col">#</th>
-      <th scope="col">Interceptor Class</th>
-      <th scope="col">Priority</th>
-      <th scope="col">Bindings</th>
-      <th scope="col">Interception Types</th>
-    </tr>
-   </thead>
-   <tbody>
-    {#for interceptor in info:devBeanInfos.interceptors}
-    <tr>
-      <td>{interceptor_count}.</td>
-      <td><span class="class-candidate">{interceptor.interceptorClass}</span></td>
-      <td><span class="badge rounded-pill bg-info text-light larger-badge" title="Priority: {interceptor.priority}">{interceptor.priority}</span></td>
-      <td>
-       {#for b in interceptor.bindings}
-        <span class="annotation" title="Binding: {b}">{b.simpleName}</span><br/>
-       {/for}
-      </td>
-      <td>
-       <ul>
-       {#each interceptor.intercepts}
-       <li>{#interception-type it.key /} {interceptor.interceptorClass.simpleName}#{it.value.name}()</li>
-       {/each}
-       </ul>
-      </td>
-    {/for}
-   </tbody>
-  </table>
-  {/body}
-{/include}
