--- extensions/arc/deployment/src/main/resources/dev-templates/removed-beans.html
@@ -1,92 +1,0 @@
-{#include main}
-  {#style}
-   .annotation {
-     color: gray;
-     font-style: italic;
-   }
-   span.larger-badge {
-     font-size: 0.9em;
-   }
-  {/style}
-  {#title}Removed Components{/title}
-  {#body}
-  <table class="table table-striped caption-top">
-   <caption>Removed Beans</caption>
-   <thead class="thead-dark">
-    <tr>
-      <th scope="col">#</th>
-      <th scope="col">Bean</th>
-      <th scope="col">Kind</th>
-    </tr>
-   </thead>
-   <tbody>
-    {#for bean in info:devBeanInfos.removedBeans}
-    <tr>
-      <td>{bean_count}.</td>
-      <td>
-       {#display-bean bean/}
-      </td>
-      <td>
-       {#bean-declaration bean/}
-      </td>
-    {/for}
-   </tbody>
-  </table>
-  
-  {#if !info:devBeanInfos.removedInterceptors.empty}
-  <table class="table table-striped caption-top">
-   <caption>Removed Interceptors</caption>
-   <thead class="thead-dark">
-    <tr>
-      <th scope="col">#</th>
-      <th scope="col">Interceptor</th>
-      <th scope="col">Bindings</th>
-    </tr>
-   </thead>
-   <tbody>
-    {#for interceptor in info:devBeanInfos.removedInterceptors}
-    <tr>
-      <td>{interceptor_count}.</td>
-      <td>
-       {interceptor.interceptorClass}
-      </td>
-      <td>
-        {#for b in interceptor.bindings}
-        <span class="annotation" title="Binding: {b}">{b.simpleName}</span><br/>
-       {/for}
-      </td>
-    {/for}
-   </tbody>
-  </table>
-  {/if}
-  
-  {#if !info:devBeanInfos.removedDecorators.empty}
-  <table class="table table-striped caption-top">
-   <caption>Removed Decorators</caption>
-   <thead class="thead-dark">
-    <tr>
-      <th scope="col">#</th>
-      <th scope="col">Decorator</th>
-      <th scope="col">Delegate</th>
-    </tr>
-   </thead>
-   <tbody>
-    {#for decorator in info:devBeanInfos.removedDecorators}
-    <tr>
-      <td>{decorator_count}.</td>
-      <td>
-       {decorator.decoratorClass}
-      </td>
-      <td>
-        {#for q in decorator.delegateQualifiers}
-        <span class="annotation" title="Binding: {q}">{q.simpleName}</span><br/>
-        {/for}
-        <span class="class-candidate">{decorator.delegateType}</span>
-      </td>
-    {/for}
-   </tbody>
-  </table>
-  {/if}
-  
- {/body}
-{/include}
