--- extensions/arc/deployment/src/main/resources/dev-templates/tags/bean-declaration.html
@@ -1,10 +1,0 @@
-{#switch it.kind}
-  {#case METHOD}
-  <span class="badge badge-success larger-badge">Producer method</span><br><code><span title="{it.declaringClass}">{it.declaringClass.simpleName}</span>.{it.memberName}()</code>
-  {#case FIELD}
-  <span class="badge badge-success larger-badge">Producer field</span><br><code><span title="{it.declaringClass}">{it.declaringClass.simpleName}</span>.{it.memberName}</code>
-  {#case CLASS}
-  <span class="badge badge-primary larger-badge">Class</span>
-  {#else}
-  <span class="badge badge-secondary larger-badge">Synthetic</span>
-{/switch}
