--- extensions/arc/deployment/src/main/resources/dev-templates/tags/duration.html
@@ -1,3 +1,0 @@
-<span class="badge badge-info"{#if it.durationMillis le 0} title="{it.duration} ns"{/if}>
-  {#if it.durationMillis > 0}{it.durationMillis} ms{#else}&lt; 1 ms{/if}
-</span>
