--- extensions/arc/deployment/src/main/resources/dev-templates/tags/kind.html
@@ -1,8 +1,0 @@
-{#when it.kind}
-  {#is PRODUCER}
-<span class="badge badge-success">Producer</span>
-  {#is DISPOSER}
-<span class="badge badge-warning">Disposer</span>
-  {#is OBSERVER}
-<span class="badge badge-secondary">Observer</span>  
-{/when}
