--- extensions/arc/deployment/src/main/resources/dev-templates/tags/tree.html
@@ -1,15 +1,0 @@
-<li>
-{#if !root}
-{#if next}├─{#else}└─{/if}
-{/if}
-{#if it.children}
-<span class="caret declaring-class">{it.declaringClassName}#</span>{it.method.name}() {#duration it /} {#kind it /}{#if it.message != null} <span class="badge badge-danger" title="Exception message: {it.message}">Error</span>{/if}
-<ul class="nested">
-{#each it.children}
-{#tree it root=false next=it_hasNext /}
-{/each}
-</ul>
-{#else}
-<span class="declaring-class">{it.declaringClassName}#</span>{it.method.name}() {#duration it /} {#kind it /}{#if it.message != null} <span class="badge badge-danger" title="Exception message: {it.message}">Error</span>{/if}
-{/if}
-</li>
