+++ extensions/arc/deployment/src/test/java/io/quarkus/arc/test/devconsole/DevObserverInfoTest.java
@@ -12,8 +12,8 @@ import jakarta.enterprise.event.TransactionPhase;
 
 import org.junit.jupiter.api.Test;
 
-import io.quarkus.arc.deployment.devconsole.DevObserverInfo;
-import io.quarkus.arc.deployment.devconsole.Name;
+import io.quarkus.arc.deployment.devui.DevObserverInfo;
+import io.quarkus.arc.deployment.devui.Name;
 
 public class DevObserverInfoTest {
