--- extensions/kafka-streams/deployment/src/main/resources/dev-templates/embedded.html
@@ -1,3 +1,0 @@
-<a href="{urlbase}/kafka-streams-topology" class="badge badge-light">
-    <i class="fa fa-project-diagram fa-fw"></i>
-    Topology </a>
