--- extensions/kafka-streams/deployment/src/main/resources/dev-templates/tags/topologyNotFound.html
@@ -1,16 +1,0 @@
-<div class="row justify-content-center">
-  <div class="col-6">
-    <div class="jumbotron">
-      <h1 class="display-4">No topology found.</h1>
-      <p class="lead">
-        Verify you have an <code>@ApplicationScoped</code> bean which defines a CDI producer method with a <code>@Produces</code> annotation returning the Kafka Streams <code>Topology</code>.
-      </p>
-      <hr class="my-4">
-      <p class="lead">
-        <a class="btn btn-primary btn-lg" href="https://quarkus.io/guides/kafka-streams" role="button">
-          <i class="fa fa-book"></i> Quarkus Apache Kafka Streams guide
-        </a>
-      </p>
-    </div>
-  </div>
-</div>
