+++ extensions/vertx-http/deployment/src/main/java/io/quarkus/vertx/http/deployment/devmode/ArcDevProcessor.java
@@ -11,7 +11,6 @@ import io.quarkus.arc.InjectableBean;
 import io.quarkus.arc.deployment.ArcConfig;
 import io.quarkus.arc.deployment.ValidationPhaseBuildItem;
 import io.quarkus.arc.deployment.ValidationPhaseBuildItem.ValidationErrorBuildItem;
-import io.quarkus.arc.deployment.devconsole.ArcDevConsoleProcessor;
 import io.quarkus.arc.processor.BeanInfo;
 import io.quarkus.arc.processor.BuildExtension;
 import io.quarkus.arc.processor.DecoratorInfo;
@@ -27,6 +26,7 @@ import io.quarkus.vertx.http.deployment.RouteBuildItem;
 import io.quarkus.vertx.http.runtime.devmode.ArcDevRecorder;
 
 public class ArcDevProcessor {
+    private static final String BEAN_DEPENDENCIES = "io.quarkus.arc.beanDependencies";
 
     @Record(ExecutionTime.RUNTIME_INIT)
     @BuildStep(onlyIf = IsDevelopment.class)
@@ -72,7 +72,7 @@ public class ArcDevProcessor {
         routes.produce(nonApplicationRootPathBuildItem.routeBuilder()
                 .route(beansPath)
                 .displayOnNotFoundPage("Active CDI Beans")
-                .handler(recorder.createBeansHandler(ArcDevConsoleProcessor.BEAN_DEPENDENCIES)).build());
+                .handler(recorder.createBeansHandler(BEAN_DEPENDENCIES)).build());
 
         routes.produce(nonApplicationRootPathBuildItem.routeBuilder()
                 .route(removedBeansPath)
