--- integration-tests/devmode/src/test/java/io/quarkus/test/devconsole/DevConsoleSmokeTest.java
@@ -1,28 +1,0 @@
-package io.quarkus.test.devconsole;
-
-import org.hamcrest.Matchers;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.RegisterExtension;
-
-import io.quarkus.test.QuarkusDevModeTest;
-import io.restassured.RestAssured;
-
-/**
- * Note that this test cannot be placed under the relevant {@code -deployment} module because then the DEV UI processor would
- * not be able to locate the template resources correctly.
- */
-public class DevConsoleSmokeTest {
-
-    @RegisterExtension
-    static final QuarkusDevModeTest config = new QuarkusDevModeTest()
-            .withEmptyApplication();
-
-    @Test
-    public void testDevConsoleNotBroken() {
-        RestAssured.with()
-                .get("q/dev")
-                .then()
-                .statusCode(200).body(Matchers.containsString("Dev UI"));
-
-    }
-}
